Emu2/requirements.txt
torch
transformers==4.37.0
peft
numpy
Pillow
argparse
einops
timm
xformers
safetensors
sentencepiece==0.1.99
decord
diffusers==0.24.0
torchvision
accelerate==0.25.0
gradio==3.40.1
tokenizers==0.13.3
opencv-python
protobuf
bitsandbytes
tensorboard